switch to float

+++ src/aligner/localaligner.h
@@ -7,8 +7,8 @@
 template <class Similarity_Matrix_Type>
 class LocalAligner {
   public:
-    virtual double calculateScore() = 0;
-    virtual double getScore() const = 0;
+    virtual float calculateScore() = 0;
+    virtual float getScore() const = 0;
     virtual unsigned int getPos() const = 0;
     virtual std::string_view getConsensus_x() const = 0;
     virtual std::string_view getConsensus_y() const = 0;
@@ -19,8 +19,8 @@ class LocalAligner {
 template <class Similarity_Matrix_Type, class LocalAligner_Type>
 class ParallelLocalAligner {
   public:
-    virtual double calculateScore() = 0;
-    virtual double getScore() const = 0;
+    virtual float calculateScore() = 0;
+    virtual float getScore() const = 0;
     virtual unsigned int getPos() const = 0;
     virtual std::string_view getConsensus_x() const = 0;
     virtual std::string_view getConsensus_y() const = 0;
